Add find by Nric functionality to relevant class
